Add SetAt on Registers
