+++ crates/cpu/src/lib.rs
@@ -7,7 +7,7 @@ pub mod interface;
 pub mod nextpc;
 mod opcodes;
 pub mod registers;
-pub mod get_at;
+pub mod futures;
 pub mod reader;
 
 pub use crate::bus::RegisterBus;
@@ -15,3 +15,4 @@ pub use crate::cpu::Cpu;
 pub use crate::interface::{NewRegisters, Registers};
 pub(crate) use crate::reader::Reader;
 pub(crate) use flags::Flags;
+pub(crate) use crate::futures::{GetAt, SetAt};
