+++ crates/cpu/src/opcodes/rotate.rs
